--- dev-server/src/runner/shared.ts
@@ -1,6 +1,0 @@
-import { FlydeFlow } from "@flyde/core";
-
-export interface FlowJob {
-    flow: FlydeFlow;
-    id: string;
-}
