+++ flow-editor/src/flow-editor/ports/ports.ts
@@ -5,8 +5,9 @@ import {
   NodeLibraryData,
   ResolvedDependenciesDefinitions,
   noop,
+  FlowJob,
+  ImportablesResult,
 } from "@flyde/core";
-import { FlowJob, ImportablesResult } from "@flyde/dev-server";
 import { ReportEvent } from "./analytics";
 import { toastMsg } from "../../toaster";
